+++ src/main/kotlin/com/knu/mockin/model/dto/request/quotations/basic/real/SearchInfoRequestParameterDto.kt
@@ -1,4 +1,4 @@
-package com.knu.mockin.model.dto.request.basic
+package com.knu.mockin.model.dto.request.quotations.basic.real
 
 data class SearchInfoRequestParameterDto (
     val prdtTypeCd: String = "", // 상품유형코드
